+++ CHANGES.md
@@ -18,6 +18,8 @@ The released versions correspond to PyPI releases.
   (see [#959](../../issues/959))
 * fixed handling of directory enumeration and search permissions under Posix systems
   (see [#960](../../issues/960))
+* fixed creation of the temp directory in the fake file system after a filesystem reset
+  (see [#965](../../issues/965))
 
 ## [Version 5.3.5](https://pypi.python.org/pypi/pyfakefs/5.3.5) (2024-01-30)
 Fixes a regression.